Cosas hechas: devuelve el vector de sites y usar fork join, falta tags y totals

+++ src/config.rs
@@ -4,7 +4,7 @@
 /// Config struct.
 /// Contains the desired number of worker threads to process the data.
 pub struct Config {
-    pub number_of_threads: u8,
+    pub number_of_threads: usize,
 }
 
 impl Config {
@@ -13,7 +13,9 @@ impl Config {
         if args.is_empty() {
             return Err("not enough arguments".to_owned());
         }
-        let number_of_threads = args[1].parse::<u8>().map_err(|error| error.to_string())?;
+        let number_of_threads = args[1]
+            .parse::<usize>()
+            .map_err(|error| error.to_string())?;
         if number_of_threads == 0 {
             return Err("Invalid number of worker threads".to_string());
         }
